refactor(go): extract l2topology model types
